route-grouping

+++ app/Http/Controllers/IdeaController.php
@@ -19,6 +19,8 @@ class IdeaController extends Controller
             'content' => 'required|min:3|max:255'
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success','Idea created successfully !');
@@ -26,7 +28,12 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
-        // where id=1;
+
+        if (auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success','Idea deleted successfully !');
@@ -34,6 +41,12 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+
+        if (auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+
         $editing = true;
 
         return view('ideas.show',compact('idea','editing'));
@@ -41,6 +54,12 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
+
+        if (auth()->id() !== $idea->user_id)
+        {
+            abort(404);
+        }
+
         $validated = request()->validate([
             'content' => 'required|min:3|max:255'
         ]);
